Extract basic editor functionality into a new class
